Add code to check return value from unlink()
This patch adds a check to the unlink() call made in util.c, which is
not currently done.  If the return value isn't zero, an error in
unlinking or removing the file has occurred.  This new code adds the
proper check, though I don't know if you want to add a different return
value in case unlink() fails.